requirements.txt
psycopg2>=2.8.6
psycopg2-binary==2.9.2
Django==3.2.25
djangorestframework==3.13.1
djangorestframework-simplejwt